+++ python/refrase/models/mistral.py
@@ -7,9 +7,9 @@ Consolidates 4 separate CareerCatalyst adapters into one file:
 - Ministral (14B, 8B, 3B)
 """
 
-from fwip.helpers import add_json_reinforcement, simplify_prompt
-from fwip.models._base import BaseAdapter
-from fwip.types import Change, ModelFamily, ModelInfo, TaskType
+from refrase.helpers import add_json_reinforcement, simplify_prompt
+from refrase.models._base import BaseAdapter
+from refrase.types import Change, ModelFamily, ModelInfo, TaskType
 
 
 class MistralAdapter(BaseAdapter):